Change as an off_t

+++ sysdeps/nacl/irt_syscalls.c
@@ -398,7 +398,7 @@ int (*__nacl_irt_fchmod) (int fd, mode_t mode);
 int (*__nacl_irt_fchdir) (int fd);
 int (*__nacl_irt_fsync) (int fd);
 int (*__nacl_irt_fdatasync) (int fd);
-int (*__nacl_irt_sync_file_range) (int fd, long long offset, long long nbytes, unsigned int flags); 
+int (*__nacl_irt_sync_file_range) (int fd, off_t offset, off_t nbytes, unsigned int flags); 
 int (*__nacl_irt_getuid) (void);
 int (*__nacl_irt_geteuid) (void);
 int (*__nacl_irt_getgid) (void);
@@ -589,7 +589,7 @@ static int nacl_irt_fdatasync (int fd)
 	return NACL_SYSCALL (fdatasync) (fd);
 }
 
-static int nacl_irt_sync_file_range (int fd, long long offset, long long nbytes, unsigned int flags)
+static int nacl_irt_sync_file_range (int fd, off_t offset, off_t nbytes, unsigned int flags)
 {
 	return NACL_SYSCALL (sync_file_range) (fd, offset, nbytes, flags);
 }
